Common implementation for `find_first_set()`
